Authentication and Authorization

+++ controllers/api/v1/posts_api.js
@@ -26,6 +26,7 @@ module.exports.destroy = async function(req,res){
     try{
         let post= await Post.findById(req.params.id);
         
+        if(post.user == req.user.id){
                 post.remove();
                 
                 await Comment.deleteMany({post: req.params.id});
@@ -33,16 +34,18 @@ module.exports.destroy = async function(req,res){
                 return res.json(200, {
                     message: "Post & associated Comments deleted successfully!"
                 });
-                
-        // } else{
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
-    }catch(err){
+            } 
+            else{
+                return res.json(401, {
+                    message: "You cannot delete this post!"
+                })
+            }
+        } 
+    catch(err){
         
         console.log('*****', err);
         return res.json(500, {
             message: "Internal Server Error"
         });
     }
-    }
+}
